Unify IndexBuffer & VertexBuffer headers for all APIs. Move API-independent parts of vertex / index buffers to their own implementation files to eliminate code duplication.
